Move the alert backtest onto the rule itself
