+++ Sources/ScoutUI/Monitoring/Alerts/Rule/AlertRule+Backtest.swift
@@ -7,12 +7,21 @@
 
 import Foundation
 
-struct AlertBacktest: Equatable {
-    static let windowBuckets = 24
+extension AlertRule {
+    private static let windowBuckets = 24
 
-    let values: [Double]
+    func backtestSummary(over values: [Double]) -> String {
+        switch backtestFireCount(over: values) {
+        case 0:
+            "Would not have fired in the past week"
+        case 1:
+            "Would have fired once in the past week"
+        case let count:
+            "Would have fired \(count) times in the past week"
+        }
+    }
 
-    func fireCount(for rule: AlertRule) -> Int {
+    func backtestFireCount(over values: [Double]) -> Int {
         let window = Self.windowBuckets
         let evaluator = AlertEvaluator()
 
@@ -28,7 +37,7 @@ struct AlertBacktest: Equatable {
                 baseline: Array(values[(end - window * 2)..<(end - window)]).median ?? 0,
                 recent: Array(values[(end - window)..<end])
             )
-            let outcome = evaluator.evaluate(rule, reading: reading, state: state, now: Date(timeIntervalSince1970: 0))
+            let outcome = evaluator.evaluate(self, reading: reading, state: state, now: Date(timeIntervalSince1970: 0))
 
             state = outcome.state
 
@@ -38,16 +47,5 @@ struct AlertBacktest: Equatable {
         }
 
         return fires
-    }
-
-    func summary(for rule: AlertRule) -> String {
-        switch fireCount(for: rule) {
-        case 0:
-            "Would not have fired in the past week"
-        case 1:
-            "Would have fired once in the past week"
-        case let count:
-            "Would have fired \(count) times in the past week"
-        }
     }
 }
